Cerrados protocolos de Estacionamientos.

+++ src/ar/edu/unq/compras/CompraPuntual.java
@@ -7,19 +7,19 @@ import ar.edu.unq.zona.PuntoDeVenta;
 
 public class CompraPuntual extends Compra {
 
-	private double horasCompradas;
+	private Integer horasCompradas;
 	
 	
-	public CompraPuntual(Integer numeroC, LocalTime horaEfectuada, LocalDate fechaEfectuada, PuntoDeVenta puntoDeVenta, double cantidadHorasCompradas) {
+	public CompraPuntual(Integer numeroC, LocalTime horaEfectuada, LocalDate fechaEfectuada, PuntoDeVenta puntoDeVenta, Integer cantidadHorasCompradas) {
 		super(numeroC, horaEfectuada, fechaEfectuada, puntoDeVenta);
 		this.setHorasCompradas(cantidadHorasCompradas);
 	}
 	
-	public double getHorasCompradas() {
+	public Integer getHorasCompradas() {
 		return horasCompradas;
 	}
 	
-	public void setHorasCompradas(double horasCompradas) {
+	public void setHorasCompradas(Integer horasCompradas) {
 		this.horasCompradas = horasCompradas;
 	}
